pyproject.toml
[project]
name = "prediction-tracker-api"
version = "0.1.0"
description = "Prediction Lifecycle Management System API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "google-auth>=2.36.0",
    "PyJWT>=2.10.0",
    "httpx>=0.28.0",
    "bcrypt>=4.2.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.17",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
